Rename test data, move to test package
