+++ api/src/test/scala/hmda/api/http/ProcessingHttpApiSpec.scala
@@ -5,7 +5,7 @@ import akka.http.scaladsl.model.StatusCodes
 import akka.http.scaladsl.testkit.ScalatestRouteTest
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
 import akka.util.Timeout
-import hmda.api.demo.DemoData
+import hmda.api.util.TestData
 import hmda.api.model.processing.{ Institution, Institutions }
 import hmda.api.processing.submission.InstitutionsFiling._
 import org.scalatest.{ BeforeAndAfterAll, MustMatchers, WordSpec }
@@ -21,7 +21,7 @@ class ProcessingHttpApiSpec extends WordSpec with MustMatchers with ScalatestRou
 
   override def beforeAll(): Unit = {
     createInstitutionsFiling(system)
-    DemoData.loadData(system)
+    TestData.loadData(system)
   }
 
   "Processing HTTP service" must {
@@ -37,7 +37,7 @@ class ProcessingHttpApiSpec extends WordSpec with MustMatchers with ScalatestRou
 
       Get("/institutions/period/2017") ~> processingRoutes ~> check {
         status mustBe StatusCodes.OK
-        responseAs[Institutions] mustBe (Institutions(DemoData.institutions))
+        responseAs[Institutions] mustBe (Institutions(TestData.institutions))
       }
     }
   }
